backend/requirements.txt
altair==5.0.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==23.1.0
billiard==4.2.1
blinker==1.6.2
cachetools==5.3.1
celery==5.5.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
fastapi==0.115.12
gitdb==4.0.10
GitPython==3.1.32
griffe==1.7.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jiter==0.9.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mcp==1.6.0
mdurl==0.1.2
numpy==1.25.2
openai==1.73.0
openai-agents==0.0.9
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pip-autoremove==0.10.0
prompt_toolkit==3.0.50
protobuf==4.23.4
pyarrow==12.0.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
redis==5.2.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.2
streamlit==1.25.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==4.3.1
urllib3==2.0.4
uvicorn==0.34.1
validators==0.20.0
vine==5.1.0
watchdog==3.0.0
wcwidth==0.2.13
zipp==3.16.2
